+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Fix excalidraw note content not copied
 * Enable (basic) spellcheck for additional languages
+* Add filters from project/design/template name in list views
 
 
 ## v2026.50 - 2026-06-17